# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Datablitzer API - OpenAPI 3.0
  description: |-
    Datablitzer api has been built for providing useful information & SEO data for specific urls and domains.

        API's include -> [authentication] -> [user] -> [url sniffing] -> [payment]
  contact:
    email: nabeel.asif362@gmail.com
  version: "1.0"
servers:
  - url: http://127.0.0.1:5001/v1
    description: Development Server
  - url: http://104.161.48.70:5001/v1
    description: Production Server
...
paths:
...
  /auth/email-verification/{token}:
    post:
      tags:
        - auth
      summary: Email verification
      operationId: emailVerification
      parameters:
        - name: token
          in: path
          description: token for email verification
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationRequest"
        required: true
      responses:
        "200":
          description: Successful operation
          # content:
          #   application/json:
          #     schema:
          #       $ref: '#/components/schemas/SignUpRequest'
        "400":
          description: Invalid code
        "500":
          description: Something went wrong, Please try again Later.
      deprecated: false
  /auth/forgot-password:
    post:
      tags:
        - auth
      summary: Forgot password
      description: provide your email to forgot password
      operationId: forgotPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: Successful operation
          # content:
          #   application/json:
          #     schema:
          #       $ref: "#/components/schemas/SignUpRequest"
        "404":
          description: User not found
        "500":
          description: Something went wrong, Please try again later.
      deprecated: false
...
  /sniffer/url:
    post:
      tags:
        - url sniffer
      summary: Add urls
      operationId: AddUrls
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddUrlsRequest"
        required: true
      responses:
        "200":
          description: successful operation
          # content:
          #   application/json:
          #     schema:
          #       $ref: '#/components/schemas/Pet'
          headers: {}
        "402":
          description: You do not have enough credits for this request, please buy more credits to perform this action
        "500":
          description: Something went wrong, Please try again later
      deprecated: false
      security:
        - bearer: []
...
components:
  schemas:
    SignUpRequest:
      title: SignUpRequest
      required:
        - first_name
        - last_name
        - email
        - password
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        first_name: super
        last_name: striker
        email: superstriker707@gmail.com
        password: test1234
...
    ForgotPasswordRequest:
      title: ForgotPasswordRequest
      required:
        - email
      type: object
      properties:
        email:
          type: string
      example:
        email: superstriker707@gmail.com
...
    EmailVerificationRequest:
      title: EmailVerificationRequest
      required:
        - code
      type: object
      properties:
        code:
          type: string
      example:
        code: "720722"
...
    AddUrlsRequest:
      title: AddUrlsRequest
      required:
        - urlArr
        - rrule
        - currIndex
        - controlSignal
      type: object
      properties:
        urlArr:
          type: array
          items:
            $ref: "#/components/schemas/UrlArr"
          description: ""
        rrule:
          type: string
        currIndex:
          type: integer
          format: int32
        controlSignal:
          $ref: "#/components/schemas/ControlSignal"
      example:
        urlArr:
          - url: https://www.google.com/
            depth: 0
            queueId: test12345
          - url: https://www.linkedin.com/
            depth: 0
            queueId: test12345
        rrule: "Test Example 1"
        currIndex: 0
        controlSignal:
          stop: false
...
    UrlArr:
      title: UrlArr
      required:
        - url
        - depth
        - queueId
      type: object
      properties:
        url:
          type: string
        depth:
          type: integer
          format: int32
        queueId:
          type: string
      example:
        url: https://www.google.com/
        depth: 0
        queueId: test12345
    ControlSignal:
      title: ControlSignal
      required:
        - stop
      type: object
      properties:
        stop:
          type: boolean
      example:
        stop: false
  securitySchemes:
    bearer:
      type: http
      scheme: bearer